introduce aggregating pool interface
